+++ interfaces/src/main/java/ar/edu/itba/paw/persistance/ImageDao.java
@@ -3,11 +3,11 @@ package ar.edu.itba.paw.persistance;
 import java.util.Optional;
 
 public interface ImageDao {
-    long create(byte[] bytes);
+    int create(byte[] bytes);
 
-    Optional<byte[]> getById(long imageId);
+    Optional<byte[]> getById(int imageId);
 
-    boolean update(long imageId, byte[] bytes);
+    boolean update(int imageId, byte[] bytes);
 
-    boolean delete(long imageId);
+    boolean delete(int imageId);
 }
