+++ interfaces/src/main/java/ar/edu/itba/paw/persistance/UserDao.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 public interface UserDao {
     User create(String username, String password, String name, String email);
 
-    Optional<User> getById(long userId);
+    Optional<User> getById(int userId);
 
     Optional<User> getByEmail(String email);
 }
